Separate the codes to several files
